// src/app/models/cliente-computador.ts
export class ClienteComputador {
    descripcion?: any;
    fecha?: any;
    horaFin?: any;
    horaInicio?: any;
    idCliente?: any;
    idInventario?: any;
   
}

export class ClienteTablaLista {

}

export class ActualizarEstado {
    estadoPrestamo?: any;
    id?: any;
}

export class HoraFin {
    horaFin?: any;
    id?: any;
}